chore: add 'gnodev build' and refactor the precompile utilities
