+++ cmd/gnodev/util.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"io/fs"
+	"strings"
+)
+
+func isGnoFile(f fs.DirEntry) bool {
+	name := f.Name()
+	return !strings.HasPrefix(name, ".") && strings.HasSuffix(name, ".gno") && !f.IsDir()
+}
